move spec code into internal
