+++ internal/spec/assertion.go
