+++ internal/spec/attestation_object.go
