+++ public_keys.go
@@ -7,7 +7,7 @@ import (
 	"crypto/x509"
 
 	"github.com/spiretechnology/go-webauthn/internal/errutil"
-	"github.com/spiretechnology/go-webauthn/spec"
+	"github.com/spiretechnology/go-webauthn/internal/spec"
 )
 
 const (
